# feat(build): Migrate to Vite build tooling with proper WordPress externalization