--- a/packages/cli/vite.config.ts
+++ b/packages/cli/vite.config.ts
@@ -1,0 +1,5 @@
+import { createKernelLibConfig } from '../../vite.config.base';
+
+export default createKernelLibConfig('@geekist/wp-kernel-cli', {
+	index: 'src/index.ts',
+});