--- a/packages/kernel/src/__tests__/integration/resource-flow.test.ts
+++ b/packages/kernel/src/__tests__/integration/resource-flow.test.ts
@@ -5,7 +5,7 @@
  * Mocks REST endpoints, verifies store state after resolvers, and validates event emission.
  */
 
-import { defineResource } from '../../resource/define.js';
+import { defineResource } from '../../resource/define';
 
 // Mock window.wp global
 interface WindowWithWp extends Window {