--- a/packages/kernel/src/error/TransportError.ts
+++ b/packages/kernel/src/error/TransportError.ts
@@ -6,8 +6,8 @@
  * @module
  */
 
-import { KernelError } from './KernelError.js';
-import type { ErrorContext, ErrorData } from './types.js';
+import { KernelError } from './KernelError';
+import type { ErrorContext, ErrorData } from './types';
 
 /**
  * Error thrown when a network/HTTP request fails