--- a/packages/ui/vite.config.ts
+++ b/packages/ui/vite.config.ts
@@ -1,0 +1,5 @@
+import { createKernelLibConfig } from '../../vite.config.base';
+
+export default createKernelLibConfig('@geekist/wp-kernel-ui', {
+	index: 'src/index.ts',
+});