… app/showcase/README.md
… app/showcase/__tests__/build-verification/peer-dependencies.test.ts
… app/showcase/__tests__/e2e/sanity.spec.ts
… app/showcase/__tests__/e2e/sprint-1-resources.spec.ts
… app/showcase/package.json
… app/showcase/webpack.config.cjs
… package.json
… packages/cli/package.json
… packages/kernel/package.json
… packages/kernel/src/__tests__/build-verification/externals.test.ts
… packages/kernel/src/__tests__/build-verification/structure.test.ts
… packages/kernel/src/__tests__/build-verification/types.test.ts
… packages/kernel/src/error/__tests__/TransportError.test.ts
… packages/kernel/src/http/__tests__/fetch.test.ts
… packages/kernel/src/resource/__tests__/cache/edge-cases.test.ts
… packages/kernel/src/resource/__tests__/cache/interpolation.test.ts
… packages/kernel/src/resource/__tests__/cache/invalidate-all.test.ts
… packages/kernel/src/resource/__tests__/cache/invalidate.test.ts
… packages/kernel/src/resource/__tests__/cache/keys.test.ts
… packages/kernel/src/resource/__tests__/define-grouped-getters.test.ts
… packages/kernel/src/resource/__tests__/define-hooks.test.ts
… packages/kernel/src/resource/__tests__/define-prefetch.test.ts
… packages/kernel/src/resource/__tests__/grouped-api-factories.test.ts
… packages/kernel/src/resource/__tests__/store/actions.test.ts
… packages/kernel/src/resource/__tests__/store/cache-keys.test.ts
… packages/kernel/src/resource/__tests__/store/creation.test.ts
… packages/kernel/src/resource/__tests__/store/grouped-api.test.ts
… packages/kernel/src/resource/__tests__/store/reducer.test.ts
… packages/kernel/src/resource/__tests__/store/resolvers.test.ts
… packages/kernel/src/resource/__tests__/store/selectors.test.ts